Move `toHaveJoined` to a setup folder
This will allow us to use it elsewhere
