+++ package.json
@@ -123,6 +123,9 @@
       "**/*.(t|j)s"
     ],
     "coverageDirectory": "../coverage",
-    "testEnvironment": "node"
+    "testEnvironment": "node",
+    "setupFilesAfterEnv": [
+      "./testutils/jest.setup.ts"
+    ]
   }
 }
